requirements.txt
numpy
numpy-stl
pandas
openpyxl
plotly
streamlit
